requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
python-multipart==0.0.9
